Add generic type for NodeGroup from incoming data
This PR allows typescript users to directly have the type of the incoming data array inferred in all the following functions:
keyAccessor, start, enter, update and leave.
If this PR is accepted, nothing will change from the user's perspective except that they will enjoy code completion where they might expect it to be.